+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+- Generate an additional grouped header copy with virtual functions, data members, and static members collected together.
+
 ## [0.1.0] - 2026-05-16
 
 ### Added